fix portable types
