--- packages/db/src/schema/index.ts
@@ -1,2 +1,0 @@
-export * from "./auth";
-export * from "./post";
